Moving the HorizonUI.IntroBlock into Horizon itself
