--- packages/HorizonUI/Sources/HorizonUI/Resources/Localizable.xcstrings
@@ -1,29 +1,0 @@
-{
-  "sourceLanguage" : "en",
-  "strings" : {
-    "attempts_allowed" : {
-      "extractionState" : "manual",
-      "localizations" : {
-        "en" : {
-          "variations" : {
-            "plural" : {
-              "one" : {
-                "stringUnit" : {
-                  "state" : "translated",
-                  "value" : "%d Attempt Allowed"
-                }
-              },
-              "other" : {
-                "stringUnit" : {
-                  "state" : "translated",
-                  "value" : "%d Attempts Allowed"
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  },
-  "version" : "1.0"
-}
